island(keystone #5, step 2): move the CloudClient singleton into island + toolkit.session
The cloud-client/auth keystone — pulls backend comms ownership into island.

- island/src/services/CloudClientService.ts owns the @mentra/cloud-client
  singleton. It builds the client from island-owned pieces (the native UDP
  socket, the MMKV secure store, the cloud-status store) + the two things the
  host provides through the toolkit FRONT DOOR (not a configureRuntime adapter):
  auth.getSubjectToken (the one permanent seam) and the resolved cloud endpoints
  + LC3 frame size via toolkit.configure({config}). On init it SELF-WIRES the
  runtime cloud/cloudConnection hooks, so the host no longer injects a
  CloudRuntimeAdapter.
- toolkit.session exposes the live-session status read-model (status +
  audioTransport) + isConnected. Account ops (delete/export) stay host RestComms
  for now (cloud-client core has none) — they land with the account domain.
- Host @/services/cloudClient becomes a thin wrapper: keeps the dev/settings
  endpoint resolution (resolvedEndpoints/lc3FrameSizeBytes/cloudConfigValues,
  host-coupled) and delegates managed-photo/stream/connection to island, so
  PhonePhotoCoordinator / cloudStreamApi / the dev Cloud-URL switcher are
  untouched.
- MantleManager passes config via configure() and drops the host-injected
  cloud/cloudConnection hooks.
- bootstrap IslandConfigValues gains audioFrameSizeBytes; jest @mentra/island
  mock gains cloudClientService + toolkit.session.

Per OS-1622 this is the adapter-free path: storage/status/client all owned by
island; the configureRuntime cloud hook is now island-self-wired (deleted in the

Typecheck + full jest (281 tests) green. Device-level audio/transcription/photo
verification is still required on hardware.
