+++ mobile/modules/island/src/island.ts
@@ -12,6 +12,7 @@ import {configure, start, stop} from "./runtime/bootstrap"
 import {glasses} from "./facades/glasses"
 import {displayMirror} from "./facades/displayMirror"
 import {speech} from "./facades/speech"
+import {session} from "./facades/session"
 import {useGlassesStore} from "./stores/glasses"
 import {useDisplayStore} from "./stores/display"
 import {useCoreStore} from "./stores/core"
@@ -26,6 +27,8 @@ export const toolkit = {
   stop,
   glasses,
   speech,
+  /** Cloud (cloud-v2) live-session status — island owns the client (keystone #5). */
+  session,
   display: {
     mirror: displayMirror,
   },
